Use test fixture introduced for shared.c for libappimage tests, too
